+++ test/framework/src/main/resources/mockito-extensions/org.mockito.plugins.AnnotationEngine
@@ -1,0 +1,1 @@
+org.elasticsearch.test.mockito.SecureAnnotationEngine
