+++ test/framework/src/main/resources/mockito-extensions/org.mockito.plugins.MockMaker
@@ -1,1 +1,1 @@
-org.elasticsearch.test.SecureMockMaker
+org.elasticsearch.test.mockito.SecureMockMaker
